Revert "Update linux-x86 Go prebuilts from ab/16390839"
Revert submission 42313522-update-go-16390839

Reason for revert: Likely culprit for b/565124023  - verifying through ABTD before revert submission. This is part of the standard investigation process, and does not mean your CL will be reverted.

Bug: 565124023

Reverted changes: /q/submissionid:42313522-update-go-16390839

+++ src/runtime/utf8.go
@@ -57,10 +57,10 @@ func countrunes(s string) int {
 // If the string appears to be incomplete or decoding problems
 // are encountered (runeerror, k + 1) is returned to ensure
 // progress when decoderune is used to iterate over a string.
-func decoderune(s string, k uint) (r rune, pos uint) {
+func decoderune(s string, k int) (r rune, pos int) {
 	pos = k
 
-	if k >= uint(len(s)) {
+	if k >= len(s) {
 		return runeError, k + 1
 	}
